moved spec_utils tests out of test_utils.py and into test_args_spec.py #617.
