+++ tests/test_args_specs.py
@@ -45,19 +45,7 @@ valid_nested_types = {
 
 
 class ValidateArgsSpecs(unittest.TestCase):
-
-    def test_model_specs_serialize(self):
-        """ARGS_SPEC: test each ARGS_SPEC can serialize to JSON."""
-        from natcap.invest import spec_utils
-
-        for model_name, metadata in cli._MODEL_UIS.items():
-            model = importlib.import_module(metadata.pyname)
-            try:
-                _ = spec_utils.serialize_args_spec(model.ARGS_SPEC)
-            except TypeError:
-                self.fail(
-                    f'Failed to avoid TypeError when serializing '
-                    f'{metadata.pyname}.ARGS_SPEC')
+    """Validate the contract for patterns and types in ARGS_SPEC."""
 
     def test_model_specs_are_valid(self):
         """ARGS_SPEC: test each arg meets the expected pattern."""
@@ -325,6 +313,43 @@ class ValidateArgsSpecs(unittest.TestCase):
             self.assertTrue(letter in valid_letters)
             # should only have a letter once
             valid_letters.remove(letter)
+
+    def test_model_specs_serialize(self):
+        """ARGS_SPEC: test each ARGS_SPEC can serialize to JSON."""
+        from natcap.invest import spec_utils
+
+        for model_name, metadata in cli._MODEL_UIS.items():
+            model = importlib.import_module(metadata.pyname)
+            try:
+                _ = spec_utils.serialize_args_spec(model.ARGS_SPEC)
+            except TypeError:
+                self.fail(
+                    f'Failed to avoid TypeError when serializing '
+                    f'{metadata.pyname}.ARGS_SPEC')
+
+
+class SpecUtilsTests(unittest.TestCase):
+    """Tests for natcap.invest.spec_utils."""
+
+    def test_format_unit(self):
+        """spec_utils: test converting units to strings with format_unit."""
+        from natcap.invest import spec_utils
+        for unit_name, expected in [
+                ('meter', 'm'),
+                ('meter / second', 'm/s'),
+                ('foot * mm', 'ft · mm'),
+                ('t * hr * ha / ha / MJ / mm', 't · h · ha / (ha · MJ · mm)'),
+                ('mm^3 / year', 'mm³/yr')
+        ]:
+            unit = spec_utils.u.Unit(unit_name)
+            actual = spec_utils.format_unit(unit)
+            self.assertEqual(expected, actual)
+
+    def test_format_unit_raises_error(self):
+        """spec_utils: format_unit raises TypeError if not a pint.Unit."""
+        from natcap.invest import spec_utils
+        with self.assertRaises(TypeError):
+            spec_utils.format_unit({})
 
 
 if __name__ == '__main__':
